+++ app/preprocessing/renaming.py
@@ -13,14 +13,12 @@ from openqasm3.ast import (
 )
 from openqasm3.visitor import QASMTransformer
 
+from app.lib.ast_utils import annotate
 from app.model.SectionInfo import SectionInfo
-from app.preprocessing.utils import annotate
 
 
 class RenameRegisterTransformer(QASMTransformer[SectionInfo]):
-    """
-    Renames all declarations inside a qasm program to prevent collisions when merging.
-    """
+    """Renames all declarations inside a qasm program to prevent collisions when merging."""
 
     renames: dict[str, Identifier]
 
@@ -28,17 +26,17 @@ class RenameRegisterTransformer(QASMTransformer[SectionInfo]):
         self.renames = {}
 
     def new_identifier(
-        self, old_identifier: Identifier, context: SectionInfo
+        self,
+        old_identifier: Identifier,
+        context: SectionInfo,
     ) -> Identifier:
-        """
-        Generates a new identifier that will be globally unique even after merging multiple programs.
+        """Generates a new identifier that will be globally unique even after merging multiple programs.
         Adds the old identifier to the list of renames.
 
         :param old_identifier: The old identifier to be renamed.
         :param context: The context of the current program.
         :return: A new globally unique identifier.
         """
-
         if self.renames.get(old_identifier.name) is not None:
             raise Exception("Variable already defined")
 
@@ -49,7 +47,9 @@ class RenameRegisterTransformer(QASMTransformer[SectionInfo]):
         return new_identifier
 
     def visit_ConstantDeclaration(
-        self, node: ConstantDeclaration, context: SectionInfo
+        self,
+        node: ConstantDeclaration,
+        context: SectionInfo,
     ) -> QASMNode:
         identifier = self.new_identifier(node.identifier, context)
         return self.generic_visit(
@@ -61,23 +61,31 @@ class RenameRegisterTransformer(QASMTransformer[SectionInfo]):
         )
 
     def visit_AliasStatement(
-        self, node: AliasStatement, context: SectionInfo
+        self,
+        node: AliasStatement,
+        context: SectionInfo,
     ) -> QASMNode:
         identifier = self.new_identifier(node.target, context)
         return self.generic_visit(
-            annotate(AliasStatement(identifier, node.value), node.annotations), context
+            annotate(AliasStatement(identifier, node.value), node.annotations),
+            context,
         )
 
     def visit_QubitDeclaration(
-        self, node: QubitDeclaration, context: SectionInfo
+        self,
+        node: QubitDeclaration,
+        context: SectionInfo,
     ) -> QASMNode:
         identifier = self.new_identifier(node.qubit, context)
         return self.generic_visit(
-            annotate(QubitDeclaration(identifier, node.size), node.annotations), context
+            annotate(QubitDeclaration(identifier, node.size), node.annotations),
+            context,
         )
 
     def visit_QuantumGateDefinition(
-        self, node: QuantumGateDefinition, context: SectionInfo
+        self,
+        node: QuantumGateDefinition,
+        context: SectionInfo,
     ) -> QASMNode:
         name = self.new_identifier(node.name, context)
         return self.generic_visit(
@@ -89,7 +97,9 @@ class RenameRegisterTransformer(QASMTransformer[SectionInfo]):
         )
 
     def visit_ExternDeclaration(
-        self, node: ExternDeclaration, context: SectionInfo
+        self,
+        node: ExternDeclaration,
+        context: SectionInfo,
     ) -> QASMNode:
         identifier = self.new_identifier(node.name, context)
         return self.generic_visit(
@@ -101,7 +111,9 @@ class RenameRegisterTransformer(QASMTransformer[SectionInfo]):
         )
 
     def visit_ClassicalDeclaration(
-        self, node: ClassicalDeclaration, context: SectionInfo
+        self,
+        node: ClassicalDeclaration,
+        context: SectionInfo,
     ) -> QASMNode:
         identifier = self.new_identifier(node.identifier, context)
         return self.generic_visit(
@@ -113,21 +125,30 @@ class RenameRegisterTransformer(QASMTransformer[SectionInfo]):
         )
 
     def visit_IODeclaration(
-        self, node: IODeclaration, context: SectionInfo
+        self,
+        node: IODeclaration,
+        context: SectionInfo,
     ) -> IODeclaration:
         identifier = self.new_identifier(node.identifier, context)
         return annotate(
-            IODeclaration(node.io_identifier, node.type, identifier), node.annotations
+            IODeclaration(node.io_identifier, node.type, identifier),
+            node.annotations,
         )
 
     def visit_CalibrationDefinition(
-        self, node: CalibrationDefinition, context: SectionInfo
+        self,
+        node: CalibrationDefinition,
+        context: SectionInfo,
     ) -> QASMNode:
         name = self.new_identifier(node.name, context)
         return self.generic_visit(
             annotate(
                 CalibrationDefinition(
-                    name, node.arguments, node.qubits, node.return_type, node.body
+                    name,
+                    node.arguments,
+                    node.qubits,
+                    node.return_type,
+                    node.body,
                 ),
                 node.annotations,
             ),
@@ -135,7 +156,9 @@ class RenameRegisterTransformer(QASMTransformer[SectionInfo]):
         )
 
     def visit_SubroutineDefinition(
-        self, node: SubroutineDefinition, context: SectionInfo
+        self,
+        node: SubroutineDefinition,
+        context: SectionInfo,
     ) -> QASMNode:
         name = self.new_identifier(node.name, context)
         return self.generic_visit(
@@ -150,14 +173,12 @@ class RenameRegisterTransformer(QASMTransformer[SectionInfo]):
     # => No collisions with other blocks but theoretically without renamed globals
 
     def visit_Identifier(self, node: Identifier, _context: SectionInfo) -> Identifier:
-        """
-        Renames identifiers using the old declaration names.
+        """Renames identifiers using the old declaration names.
         ToDo: We currently ignore scope!!
 
         :param node: Identifier to rename
         :return: Modified identifier
         """
-
         new_identifier = self.renames.get(node.name)
         if new_identifier is None:
             return node
